投稿の記述をさらにコンポーネントで分けた
